split memory cell up into separate rules
